clojure.lang.APersistentMap-like hashing for maps
Introduces a new function -- cljs.core/hash-imap -- which calculates a
hash for an IMap using the algorithm from c.l.APM, which is
independent of the ordering on map entries.

This is necessary so that the various map types hash to the same value
when equal, independent of insertion ordering and the like and without
sorting of keys.