CLJS-235: implementing David Nolen's protocol mask idea
The core protocols are now partitioned into groups of 32. For each
such group a property cljs$lang$protocol_mask$partitionX$ (with X
being the partition number) is attached to *instances* (not
prototypes) of implementing types / records at construction time (in
the ctor). This is then used by satisfies? to produce quick answers in
the "success" case (that is, when the given object does satisfy a core
protocol).