CLJS-129: Fix browser-connected REPL namespace loading
The browser-connected REPL keeps track of all of the namespaces it has
loaded and will only load them once. Some namespaces will have already
been loaded when the host page was loaded in the browser. This fixes
how we attempt to determine what has already been loaded. Instead of
asking the browser what has been loaded, which is not reliable, we
allow namespaces to be loaded a second time. If loading a namespace a
second time will cause problems, the namespace can be passed to
repl-env to be ignored. See doc string for cljs.repl.browser/repl-env.