CLJS-272: support :refer and skipping :as in :require
The :as alias part of a :require spec can now be omitted.
Alternatively, a symbol naming the namespace to be required without an
alias can be specified without wrapping it in a vector.

Additionally, :require accepts a :refer option with :use-like effect.

(ns foo.core
  (:require
    ;; bar-fn will be available unqualified
    [lib.bar :as bar :refer [bar-fn]]
    ;; :as lib.baz is implicit
    [lib.baz]
    ;; likewise
    lib.quux))

Either, both or none of :as, :refer may be specified in each :require
spec.

:require-macros supports the same options.

:use / :use-macros is now supported by rewriting the specs as :require
/ :require-macros specs with :refer.

:require / :require-macros now produce problem-specific error
messages which include the offending spec.

Tests are provided for :refer and the new handling of :use.