CLJS-403 REPL is changed to analyze forms in an :expr context.
Thereby making explicit, that the REPL actually compiles each input in an :expr context; in a wrapper form, to capture the return value. When an exception is printed, the REPL attaches the js compiled from the original.

Several builtin forms, however, don't get emitted in a :statement context, which could lead to apparent inconsistencies when looking at the js on a stacktrace.