CLJS-491: avoid layered subvecs
cljs.core/build-subvec now checks whether its vector argument is
already a Subvec and, if so, unwraps it, so that the return value of
cljs.core/subvec is always a view on a regular vector. This is in line
with the behaviour of subvec in Clojure.

Includes a basic test.