Send inline source map information for form evaluation
When evaluating forms from a browser repl it is nice to see and get errors
referencing source information. To achieve that:

1. Send a sourceMappingURL with a data url using a base64 encoded
source map
2. Include sourcesContent in the source map based on the form being evaluated
3. Use sourceURL to name the eval block as the generated file
referenced in the source map.

This patch uses the :source metadata from tools.reader as the value to send
back to the browser when possible.  In the case of something without metadata
(string, number, etc) it will use (pr) to get the string representation.

Source maps are only generated and sent for a repl-env with a truthy
:source-map field, so the browser-repl has been changed to merge that value.
Make work