CLJS-858: resolve-existing-var does not check vars outside current namespace
Prior to this commit we would only check vars if the prefix of the
var was the current namespace. This means normal library usage
is not verified. Instead now we check if the var is either:

A) an implicit require/import, i.e. goog.string, Math, etc.
B) is from a JS namespace (just verify the namespace)
C) exists in a CLJS namespace