Add static support for vars
Add `Var` type, supports `IMeta`, `IDeref`,
`IPrintWithWriter`. metadata includes all the same information as
Clojure with the exception of `:arglists`.

Stop calling `read-string` on the results from a ClojureScript REPL
evaluation. It's not clear what problem this was attempting to solve
and it prevented the `Var` type from printing correctly.