re-introduce cljs.closure/compile-file skip analysis optimization
builds are "non-interactive". That is, the unit of compilation for projects is
nearly always at the module (namespace) level. Here ns information triggers
analysis when required so skipping analysis is not problematic. builds now
default to "non-interactive".

Cold builds when there is no work to do is now quite fast.

the real default is "interactive" so REPLs can continue to work as they have.