CLJS-1158: Regression: compiler fails to see symbols defined in another namespace
Namespaces and macro namespaces can share the same name. In
the case that some other namespace loads a macro namespace
this will prevent analysis of the runtime namespace because
`cljs.analyzer/analyze-deps` only examines the existence of a
ns entry, not the presence of actual var mappings - `:defs`. We
now check for `:defs`.

Also when writing out the analysis cache we always drop `:macros`
as we do not track macro namespaces on disk for changes. We
should always reload macros and populate `:macros` fresh.