CLJS-1301: local :foreign-libs are not picked up the first time browser REPL is started
The 2-arity version of cljs.closure/build simply adopts cljs.env/*compiler* if
there is a root binding. The browser REPL was invoking this arity of build to
generate the REPL iframe JS and thus clobbering it.

Create a fresh ambient compiler environment instead in
cljs.repl.browser/compile-client-js and invoke the 3-arity version of
cljs.closure/build