CLJS-934: In the REPL return vars after defs
Introduces a new REPL option :def-emits-vars
which defaults to true. This REPL option is
conveyed to the analysis/compilation environment
by cljs.repl. When set to true, the analyzer
includes the var AST in the def AST and the
compiler the var AST when emitting the JavaScript
for def, arranging things so that a var is returned
instead of the def init value.