CLJS-1092: CommonJS Module processing
Adds support for CommonJS modules. CommonJS modules can be included as
foreign-libraries using the new compiler option :module-type and the
value :commonjs, e.g.

:foreign-libs [{:file "libs/greeting.js"
                :provides ["greeting"]
                :module-type :commonjs}]

Foreign libraries that have a module-type will be processed before
creating the dependecy index. They will be converted into Google Closure
modules using functionality from the Google Closure compiler and will be
written to disk. We can then change the compiler options to include the
modules using the :libs option. We also need to set a mapping from the
old module name to the new module name in the compiler
environment. Later, when the sources get analyzed and compiled, we check
our mappings and replace occurences of the old module name with the new
module name.
In case the user supplies a module-type other than :commonjs, a
warning will be shown, the module will be written to disk as is and we
won't add a mapping to the compiler environment.