CLJS-1304: Behavior of c.string/replace differs from Clojure
JavaScript own's `String.prototype.replace` knows how to do string
replacement using functions, therefore is used as internal
implementation. But the signature differs as the js version feeds the
matched groups as variadic arguements to the replacement function. The
solution here is to create a "function adapter" between the js replace
and the replacement function.

For cases where we have a single matching string without any subgroup,
we pass it directly, without the wrapping vector.