introduce cljs.analyzer/*cljs-macros-ns* to track entrance into a macro ns compilation or analysis. All files related to the macro ns are themselves macro nses - i.e. a phase distinction.
in cljs.compiler/compile-file this dyn var is bound to true only if we have
been given a file with a .clj extension, cljs.core with a .cljc extenion, an
opts with a :macros-ns key with a truth-y value, or if a the dyn var has
already been bound.

in all cases where we special case cljs/core.cljs for AOT reasons,
disambiguate by checking the file extension. This way we don't treat
attempts to bootstrap the cljs/core.cljc the macro ns as an attempt
to load the standard lib.