CLJS-1317: Incremental compilation issues for :nodejs target
The issue was that the AOTed cljs.core is compiled with :target :default. This
triggered recompilation under :target :nodejs. Since core is recompiled all other
namespaces will be recompiled due to :recompile-dependents true.

Don't recompile because of differing build options if the namespace is cljs.core
and AOTed core exists.