CLJS-1449: self-host: Error if :require-macros with :as
The k argument to cljs.js/load-macros should be either
a keyword or nil. The current implementation employs
an or construct that can set k to be the value false, and
when ultimately passed back into load-macros this
false value can inadvertently be used as a function.

This patch

a) revises the or construct to terminate wil a nil,
   ensuring k will be be set to nil where previously it
   would have been set to false

b) revises a spot where k is used to look up a value
   in a map, reversing things so that the map is
   instead used as the (non-nil) fn