CLJS-1467: Foreign Libraries not included when using :main with :simple or :advanced
Fixes bug by checking if either output-file or output-to option is
set, which means that compile-file will return a IJavaScript map instead
of a String when output-to option is set. Need to check if output-to is
a String, since it also could be the :print keyword, in which case we
don't want to write the output to disk.