CLJS-1423: self-host: Requiring analyzer/compiler breaks unchecked Boolean
When analyzing, trap set! calls targeting
*unchecked-if* iff the value is a compile-time
boolean constant.

Otherwise, if the cljs.analyzer namespace
is required by ClojureScript code being
compiled by ClojureScript-JVM, when
the form

  (set! *unchecked-if* val)

in the :cljs conditional-compilation branch
is analyzed, it is unintentionally trapped and
the *unchecked-if* is reset to the symbol
'val (which is truthy, causing the compiler
to then emit unchecked if code).