CLJS-1476: Self-host: Protocol prefixing broken for three- (or more) segment namespaces
Fixes protocol-prefix fn (two copies, one in core and one
in compiler) for use in self-host. At its core, this function
need to take a string like

  alpha.beta.gamma/IFoo

and turn it into

  alpha$beta$gamma$IFoo$

and it was failing to replace the multiple dot characters
when in self-host mode owing to the fact that .replace
in JavaScript is not a global replace. This fix adds the
"g" flag.