CLJS-1506: doc for referred fn displays alias ns
If you get the var for a referred symbol, the resolved var
has the target namespace rather than the source. For
example, (require '[clojure.string :refer [trim]]) results
in (var trim) evaluating to #'cljs.user/trim (which itself
cannot be resolved) rather than #'clojure.string/trim (as
is the behavior in Clojure).

This fixes this issue in a direct way, by putting the resolved
var namespace in the data returned by
cljs.analyzer/resolve-var. This fixes the underlying var
issue and, as a consequence, fixes the doc output as
well.