CLJS-1546: cljs.core/run! does not always return nil
The cljs.core/run! function only returns nil if the proc function
passed to it returns nil, but the docstring for run! indicates that it
returns nil.

Fixes this issue by explicitly returning nil. This also makes the
ClojureScript implementation of run! match that in Clojure 1.8.

Adds unit tests for run!, copying tests from Clojure that exhibit the
bug, while also adding a test for the primary functionality.