CLJS-1551: Self-host: assert-args dormant in macros
Since self-host must call macros defined in a different compilation
stage, the assert-args macro used in if-let and other macros must be
defined in a different namespace entirely (they can't be in cljs.core
because this namespace is already being compiled as a macros namespace
at this point). Introduce a cljs.support namespace for this purpose,
and to keep things simple, also arrange to have non-bootstrap use this
namespace for the purpose of assert-args.