CLJS-1556: Invalid code emit for obj literal
Certain legal ClojureScript expresssions involving #js object literals
end up causing invalid JavaScript to be emitted. In particular, this can
occur if such an object literal is in statement context. It can also
occur if simply evaluating an object literal in expression context by
itself in a REPL.

At the root of the problem is that the opening brace associated with an
object literal should not be at the beginning of a JavaScript statement.
Otherwise, the brace may be interpreted as the start of a JavaScript
block.

This patch fixes the issue by simply wrapping all emitted object literals
in parens.