CLJS-1542: Self-host: cljs/compile-str not handling errors properly
If you use cljs.js/compile-str on a form that cannot be analyzed, then
the analysis error is wrapped, but inadvertently passed on to
compilation as an AST structure. This results in the compiler derailing
because it tries to process a nil :op.

The fix is to employ the same pattern used in cljs.js/eval-str, namely:
Wrap successful analysis in a {:value ast} map, and then additionally
check for an :error key in the result and cb early, otherwise extract
the :value and continue on to compilation.