CLJS-1565: Self-host: whitespace optimization is broken
If generating single-file (in particular :whitespace) output for a
project that makes use of cljs.js, it is possible for the concatenated
file to have cljs.js appear prior to cljs.core$macros, and thus cause
a dependency issue when loading the file. This is a consequence of the
fact that the cljs.closure.JavaScriptFile compiled output for cljs.js
doesn't specify cljs.core$macros macros as a dependency.

There are already a few places where the cljs.core$macros dependency is
patched in. This revision factors out a bit of that logic and adds it
to another needed place.