CLJS-1580: Self-host: goog.provide offsets source-maps
When goog.provide is emitted as a result of processing :ns :op, this
modification to the JavaScript offsets the lines in that file by one,
thus messing up source mapping lines, if enabled. Instead of simply
using str to generate the line, use comp/emitln, which internally does
the proper accounting with respect to source maps.