CLJS-1564: Self-host: cached macro *loaded* update
When requiring namespaces, the namespace name symbol is conjed into the
*loaded* set, with the ana/macro-ns-name being conjed in the case of a
macros namespace. (So for example, 'foo.core for a regular namespace
and 'foo.core$macros for a macros namespace.)

But, if the namespace being loaded is a macros namespace for which
the client returns cached JavaScript, then the code instead conjes the
base namespace name.

This is actually the result of a simple oversight with the changes
made for CLJS-1504: One place where *loaded* is manipulated, name is
used instead of aname.