CLJS-1577: Self-host: syntax-quote resolves on dot forms
If you try to evaluate a syntax-quoted dot form, as in `.x, by default
symbol resolution will be applied to the dot form .x, yielding /x.
Instead, we need this to yield just .x. To do this, revise cljs.js
to check for this case before delegating to cljs.analyzer/resolve-symbol.