Avoid corrupted compiler env in ana/parse-ns
parse-ns allows for running the parse in two ways: updating the global
compiler env or not updating it. This is controlled via the :restore
option. To implement this, it isolates the parsing process by binding the
compiler env to a snapshot of the compiler env when it starts. When the
parsing is done, it checks the :restore option and if it's false it
merges back parts of the resulting compiler env (the ::namespaces and
::constants keys) into the global compiler env.

The problem is that during the parsing, other compiler processes could
have legitimately updated those same keys in the global compiler env, so
this merge can occasionally result in compiler data loss.

The race condition can avoided by using a different approach: only when
:restore is true the compiler runs isolated from the global compiler
env, by taking a snapshot of the env at the beginning. When :restore is
false, no snapshot is taken: env/*compiler* binding is not modified, so
the global compiler env is mutated in-place