CLJS-1649: Possible issue with in cljs.reader or cljs.core/PersistentHashMap.
We are storing last 256 computed hash-string values in a JS object. JS objects only accept strings as keys. So its not able to distinguish between a `nil` and a `"null"` if either is present in the cache already. We are already assuming that `nil` hashes to 0. So don't check for it in the cache or try to compute its hash.