CLJS-1679: Self-host: Incorporate spec tests
Adds cljs.spec-test to the self-host test suite

To do so involves some adjustments to self-host
loading in this suite (a workaround for CLJS-1657
for some of the spec namespaces) and the need
to skip loading code for cljs.core macros namespace.

This also catches and fixes one production code
issue in the cljs.spec macros namespace: the symbol
*instrument-enabled* needs to be qualified so that
it refers to the runtime namespace.