CLJS-1694: Self-host: Port macro var inference in :refer
Employ the same general strategy used in JVM ClojureScript:

Defer checking uses until after macro processing of the namespace form
is complete, feeding missing uses to check-uses-macros, employing
a factored out common check-use-macros-inferring-missing that JVM
and bootstrapped ClojureScript can use (with this factored funtion
taking care of finding inferred macros and updating the compilation
environment).

In short, the bootstrap port of the feature is faithful to the
JVM implementation of the feature.

Add a new unit test for this feature and tweak the self-host-parity
test to avoid loading clojure.template, which otherwise results in
the compiler enviroment dynamic var being inadvertently cleared.