CLJS-1695: Self-host: Port cljs / clojure namespace aliasing
If a failed attempt is made to load a clojure.* namespace, retry with
the cljs.* version. If that load succeeds, proceed to patch the alias
map in the compiler analysis metadata state with the new namespace
symbol and also rewrite the ns ast so that it will pass subsequent
namespace analysis checks and related processing.