CLJS-1759: Errors writing transit analysis cache if parallel build
Adds a coarse-grained mutex around the calls to write out transit
analysis cache data. A result of this is that sporadic failures that
only occur when concurrently writing out analysis cache info when
compiling largeish multi-file projects with :parallel-build true is
avoided.

(Perhaps there is some thread-safety defect in the transit-clj stack
yet to be discovered / fixed, and this patch acts as a suitable
workaround until a proper solution is found.)