CLJS-1773: Self-host: Don't resolve unqualified symbols / keywords with $macros
Revise cljs.js so that unqualified symbol and :: keyword namespace
resolution strip off $macros if present on the namespace.

Also, go through all of the existing ClojureScript code that employs
explicit qualification simply for the purpose of self-hosted
compatibility, and remove it, reverting to the simpler code previously
being used.