CLJS-1797: Update aot_core to support build with MINGW on Windows
In order to get Git Bash to work with ./script/build, aot_core needed
to be updated to provide Maven's dependency:build-classpath plugin
with new fileSeparator and pathSeparator values.

FILE_SEP and PATH_SEP were added to aot_core, defaulting to unix conventions.
The uname and tr commands are used to detect MINGW, and will update FILE_SEP
and PATH_SEP if needed. Once these values are set, it was necessary to use
them to ensure that the classpath is set correctly. PATH_SEP is now used to
append src/main/clojure and src/main/cljs.