Provide more descriptive error message when invalid libspec detected
When an invalid libspec is detected while compiling, an error is thrown
which doesn't specify what the invalid libspec is. Combined with the
line/column number referring to the start of the file, this can make
it difficult to track down what is causing the problem.

This patch prints the invalid libspec in the error which should make it
much easier to find and fix the problem.