CLJS-1869: Regression importing goog.Uri
Employ a new `canonicalize-import-specs` when using the new import
macro (in lieu of `canonicalize-specs`) which is faithful to the
original REPL code in a couple of respects:

* It doesn't put bare libspec symbols into vectors
* It simply allows sequential libspecs to pass through

The first change is important because it allows `parse-import-spec`'s
handling of non sequential lib specs to be applied when forming the
import map.

As a concrete example, `parse-import-spec` will convert both '[goog Uri]
and 'goog.Uri to the import map '{Uri goog.Uri}.

Without this change, for the form (import 'goog.Uri), `parse-import-spec`
will instead be passed '[goog.Uri], which is converted to an empty
import map.

The second bullet point, which we get for free with the change to a
faithful canonicalize fn, allows, for example: (import '(goog Uri)) as
was previously possible.