Check for compilation success, and lib folder
If script/bootstrap hasn't been run, then running script/test returns
the slightly cryptic error message:

  Error: Could not find or load main class clojure.main

This patch checks if the lib folder exists and has files in it. If not
it prompts the user to run script/bootstrap and exits.

This patch also adds a check for CLJS compilation success before running
tests against the various VM engines.