CLJS-1925: Use of undeclared Var cljs.user/RegExp when extending protocol for RegExp
This is a regression caused by an improper fix to CLJS-1607. This patch reverts
the fix in CLJS-1607, adds a test for it that doesn't rely on internal details,
and properly fixes the issue by not `dissoc`ing locals from env in
`cljs.core/resolve-var`, because `specify!` creates a local binding via `let`.