CLJS-1932: Self-host: Perf regression macroexpand-check
Addresses a self-hosted perf regression obtaining
cljs.spec/macroexpand-check.

Previously for every macroexpansion, if cljs.spec was loaded, it
would build a map of all of the vars in that namespace and then
get the the var for macroexpand-check from that map.

This revision avoids building the map upon every check, instead
performing a cheaper find-ns-obj check to see if cljs.spec is loaded
and if so, performing the same logic but within a delay so as to
cache the result.