CLJS-1976: hash-map assoc stackoverflow
create-node (and possibly other internal functions) assume that hash
values collide when their values are equal. Because it is possible
to return a hash value of more than 32 bits, it is possible to have a
hash collision where hashes are unequal, causing undefined behavior.
In this case, it caused a stackoverflow during hash-map and set
assoc! and assoc.

This patch forces the hash function to truncate all hash values from
dates and -hash implementations to 32 bits.