CLJS-2139: Undeclared var regression in fn bodies
With CLJS-2066, a change was made to skip analyzing named fn method
bodies on the first analysis pass, deferring this analyis to the
second, richer pass dedicated to optmizing self calls. Since the second
pass has all warnings suppressed, this introduces a subtle regression
in that no warnings would be emitted for issues found in named function
bodies.

This fixes the issue by turning off the blanket no-warn for the second
pass. Since warnings can only be emitted when analyzing method bodies
(the analysis of parameters doesn't lead to warnings), this is
sufficient to solve the problem.