refactor handle-js-modules
remove only-required? param, not used

in Node.js context we still need to build :node-module-index which
is just the entire Node.js top level. This is necessary in REPL contexts
where compilation is always incremental. We may for example need to
analyze dependencies without compiling them - but in this case we
won't know have information about deps from node_modules up front
(in a build this would not occur since we process files in dep order).

The REPL strategy could probably be refactored to remove these issues
but that is out of scope for the coming release.