CLJS-2423: Allow custom :output-wrapper function
Expands support from 2 types to 4 types:
- function
- string (`format` compatible string interpolation)
- truthy (that is of course, neither function nor string)
- falsey

Previously, any truthy value would cause the javascript to be wrapped by some
default stringified function. For most simple applications, this would have
sufficed. In cases where the `this` context in the wrapper function is used,
something more expressive is needed to be able to assign this context.
The string and function support aims to fill this gap.